Address Optionsbleed CVE-2017-9798
Apache httpd allows remote attackers to read secret data from process memory
if the Limit directive can be set in a user's .htaccess file, or if httpd.conf
has certain misconfigurations, aka Optionsbleed. The attacker sends an
unauthenticated OPTIONS HTTP request when attempting to read secret data. This
is a use-after-free issue and thus secret data is not always sent, and the
specific data depends on many factors including configuration. Exploitation
with .htaccess can be blocked with a patch to the ap_limit_section function in
server/core.c.

Signed-off-by: Peter O'Connor <peter@solus-project.com>